updates to SampleIntegration.php
One small change. [Password] needs to be in single quotes '[Password]' because the passwords generated by Mercury can contain the $ character and this will not work in PHP with double quotes like "[Password]"

Also, "LaneID" isn't listed as one of the parameters in the current documentation but I don't know if there is another reason for it being there.